Update name; add test skeleton
