+++ test/d2l-activity-name/data/assignment1.json
@@ -1,0 +1,6 @@
+{
+	"rel":["https://assignments.api.brightspace.com/rels/assignment"],
+	"properties":{
+		"name":"Folder 1"
+	}
+}
