+++ test/d2l-activity-name/data/quizAttempt1.json
@@ -1,0 +1,5 @@
+{
+	"properties":{
+		"name":"Quiz 1"
+	}
+}
