+++ test/d2l-activity-name/data/topic1.json
@@ -1,0 +1,5 @@
+{
+	"properties":{
+		"name":"Topic 1"
+	}
+}
